# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Online Finetune API
  version: 0.1.0
paths:
...
  /api/v1/datasets/{dataset_id}/query:
    post:
      tags:
        - classify
      summary: Query Dataset Data
      operationId: query_dataset_data_api_v1_datasets__dataset_id__query_post
      parameters:
        - name: dataset_id
          in: path
          required: true
          schema:
            type: string
            title: Dataset Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryDataRequest"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Query Dataset Data Api V1 Datasets  Dataset Id  Query
                  Post
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    QueryDataRequest:
      properties:
        query_type:
          type: string
          title: Query Type
          description: "One of: annotation-stats, sample-slice, metadata-histogram,
            recent-annotations, prediction-summary, wafer-points"
        params:
          additionalProperties: true
          type: object
          title: Params
      type: object
      required:
        - query_type
      title: QueryDataRequest
      description: Structured data query from the agent.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...